requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
prometheus_client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
pytest==8.2.0
